requirements.txt
requests>=2.31.0
colorama>=0.4.6
dnspython>=2.4.0